+++ RELEASE_NOTES.md
@@ -1,6 +1,26 @@
 rest_social_auth release notes
 ==============================
 
+master
+------
+
+
+v1.1.0
+------
+- Update docs
+- Add new setting `REST_SOCIAL_LOG_AUTH_EXCEPTIONS`
+
+Issues #42
+
+v1.0.0
+------
+- Add Django 1.10 support
+- Drop Django 1.7 support
+- Add social-auth-core, social-auth-app-django dependencies
+- Drop python-social-auth dependency
+
+Issues: #33, #35, #37, #38
+
 v0.5.0
 ------
 - Handle HttpResponses returned by the pipeline